backend/src/main/java/com/arnold/customer/CustomerDaoJPAImpl.java
package com.arnold.customer;

import org.springframework.stereotype.Repository;

import java.util.List;
import java.util.Optional;

@Repository("jpa")
public class CustomerDaoJPAImpl implements CustomerDao{

    private CustomerRepository customerRepository;

    public CustomerDaoJPAImpl(CustomerRepository customerRepository) {
        this.customerRepository = customerRepository;
    }

    @Override
    public List<Customer> selectAllCustomers() {
        return customerRepository.findAll();
    }

    @Override
    public Optional<Customer> selectCustomerById(Integer id) {
        return customerRepository.findById(id);
    }

    @Override
    public void insertCustomer(Customer customer) {
        customerRepository.save(customer);
    }

    @Override
    public boolean existCustomerWithEmail(String email) {
        return customerRepository.existsCustomerByEmail(email);
    }

    @Override
    public void deleteCustomerById(Customer customer) {
        customerRepository.delete(customer);
    }

    @Override
    public void updateCustomerById(Customer customer) {
        customerRepository.save(customer);
    }
}
